+++ personality/__init__.py
@@ -1,0 +1,7 @@
+"""
+Personality module for Coda Lite.
+"""
+
+from .personality_loader import PersonalityLoader
+
+__all__ = ["PersonalityLoader"]